Revert "extracted and implemented base class"
This reverts commit fcca6c60130b4654adaab3758c21cc7b7c89890e.
