+++ src/NLog/NLog.netfx45.csproj
@@ -320,7 +320,6 @@
     <Compile Include="LogReceiverService\NLogEvents.cs" />
     <Compile Include="LogReceiverService\SoapLogReceiverClient.cs" />
     <Compile Include="LogReceiverService\StringCollection.cs" />
-    <Compile Include="LogReceiverService\WcfLogReceiverClientBase.cs" />
     <Compile Include="LogReceiverService\WcfLogReceiverTwoWayClient.cs" />
     <Compile Include="LogReceiverService\WcfLogReceiverClient.cs" />
     <Compile Include="LogReceiverService\WcfLogReceiverOneWayClient.cs" />
